extract write module
