+++ with-application-events-immutable/src/main/java/io/dddbyexamples/cqrs/write/infrastructure/package-info.java
@@ -1,0 +1,1 @@
+package io.dddbyexamples.cqrs.write.infrastructure;
