+++ with-application-events-immutable/src/test/java/io/dddbyexamples/cqrs/write/domain/CreditCardTest.java
@@ -1,4 +1,4 @@
-package io.dddbyexamples.cqrs.model;
+package io.dddbyexamples.cqrs.write.domain;
 
 import org.junit.Test;
